Implement each command in a separate file
This is a mechanical refactoring to put each command's code into a separate file. Earlier, we had an ad hoc organization where a few commands were in mod.rs and the rest in their own files. Now mod.rs contains only a few `pub mod` declarations.
