+++ payas-cli/src/commands/command.rs
@@ -1,0 +1,7 @@
+use std::time::SystemTime;
+
+use anyhow::Result;
+
+pub trait Command {
+    fn run(&self, system_start_time: Option<SystemTime>) -> Result<()>;
+}
